Got rid of some duplicate code and unified some folders
