+++ src/main/java/com/revature/DAOs/LikeDao.java
@@ -1,4 +1,4 @@
-package com.revature.DAO;
+package com.revature.DAOs;
 
 import java.util.List;
