+++ src/main/resources/data.sql
@@ -14,6 +14,29 @@
     VALUES
     (3, 'bio', 'email3', 'lots of comments', 'person', 'password', 'profileimageurl', 'username3');
 
+
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('feugiat et eros vestibulum ac est lacinia nisi venenatis tristique', 'dharriskine0@canalblog.com', 'Darrelle', 'Harriskine', 'pass', 'http://dummyimage.com/201x100.png/dddddd/000000', 'User');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('lectus pellentesque eget nunc donec quis orci eget orci vehicula condimentum curabitur', 'aertel1@newsvine.com', 'Almeria', 'Ertel', 'bIHAwpp5', 'http://dummyimage.com/120x100.png/5fa2dd/ffffff', 'aertel1');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('semper interdum mauris ullamcorper purus sit amet nulla quisque arcu libero rutrum ac lobortis vel dapibus at diam nam tristique', 'bgreathead2@google.ca', 'Blake', 'Greathead', 'ZigPFqL', 'http://dummyimage.com/231x100.png/ff4444/ffffff', 'bgreathead2');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('a ipsum integer a nibh in quis justo maecenas rhoncus aliquam lacus morbi quis tortor', 'ghasluck3@soup.io', 'George', 'Hasluck', '033GcBK', 'http://dummyimage.com/193x100.png/dddddd/000000', 'ghasluck3');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('ipsum dolor sit amet consectetuer adipiscing elit proin risus praesent lectus vestibulum quam sapien varius ut', 'uregnard4@twitpic.com', 'Ulrika', 'Regnard', 'DAnDaHK3M8T', 'http://dummyimage.com/237x100.png/dddddd/000000', 'uregnard4');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('eget massa tempor convallis nulla neque libero convallis eget eleifend luctus ultricies eu nibh quisque id justo sit amet sapien', 'rmoakes5@arstechnica.com', 'Roseann', 'Moakes', 'dXlQ9WyDp', 'http://dummyimage.com/106x100.png/ff4444/ffffff', 'rmoakes5');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('est quam pharetra magna ac consequat metus sapien ut nunc vestibulum ante ipsum primis in faucibus orci luctus et', 'ipowelee6@google.com', 'Indira', 'Powelee', 'yZ7eSkXweU4', 'http://dummyimage.com/241x100.png/5fa2dd/ffffff', 'ipowelee6');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('morbi porttitor lorem id ligula suspendisse ornare consequat lectus in', 'alongstreet7@mediafire.com', 'Armando', 'Longstreet', '3cfqbs0hn', 'http://dummyimage.com/155x100.png/5fa2dd/ffffff', 'alongstreet7');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('nisi vulputate nonummy maecenas tincidunt lacus at velit vivamus vel', 'mplank8@arstechnica.com', 'Matias', 'Plank', 'RmNKM6F4', 'http://dummyimage.com/181x100.png/5fa2dd/ffffff', 'mplank8');
+	insert into USERS (bio, email, first_name, last_name, password, profile_picture, username) 
+		values ('curae duis faucibus accumsan odio curabitur convallis duis consequat dui nec nisi volutpat', 'mbullin9@sfgate.com', 'Marcy', 'Bullin', 'pDmumlS', 'http://dummyimage.com/159x100.png/5fa2dd/ffffff', 'mbullin9');   
+
+    
 -- create 6 "posts" or 1 post with 4 comments of varying depth, and a second top level post
     -- first is a top level post! note how parent_id is null in Postgresql; in Java this value will be converted into an Integer as null, meaning it is not nested at all. top level.
     INSERT INTO public.posts
